+++ indico/modules/attachments/operations.py
@@ -9,21 +9,24 @@ from flask import session
 
 from indico.core import signals
 from indico.core.db import db
+from indico.core.permissions import update_read_permissions
 from indico.modules.attachments import logger
 from indico.modules.attachments.models.attachments import Attachment, AttachmentType
 from indico.modules.attachments.models.folders import AttachmentFolder
 
 
-def add_attachment_link(data, linked_object):
+def add_attachment_link(form, linked_object):
     """Add a link attachment to linked_object."""
-    folder = data.pop('folder', None)
+    form_data = form.data
+    folder = form_data.pop('folder', None)
     if not folder:
         folder = AttachmentFolder.get_or_create_default(linked_object=linked_object)
     assert folder.object == linked_object
     link = Attachment(user=session.user, type=AttachmentType.link, folder=folder)
-    link.populate_from_dict(data, skip={'acl', 'protected'})
-    if link.is_self_protected and 'acl' in data:
-        link.acl = data['acl']
+    link.populate_from_dict(form_data, skip={'acl', 'protected'})
     db.session.flush()
+    # To be able to log attachment ID these steps must be done after flushing
+    if link.is_self_protected and 'acl' in form_data:
+        update_read_permissions(link, form)
     logger.info('Attachment %s added by %s', link, session.user)
     signals.attachments.attachment_created.send(link, user=session.user)